requirements.txt
Django==5.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
Pillow==10.4.0
python-decouple==3.8
firebase-admin==6.5.0
python-dotenv==1.0.0
user-agents==2.2.0
deep-translator==1.11.4
pandas==2.2.3
openpyxl==3.1.2
boto3==1.34.131
django-storages==1.14.4
gunicorn==21.2.0
psycopg2-binary==2.9.10
setuptools>=45.0.0